# docs/v1/v1_swagger.yaml
definitions:
  go_shurtiner_internal_app_model.CreateLink:
    properties:
      url:
        type: string
    type: object
  go_shurtiner_internal_app_model.JobQueue:
    properties:
      completed_at:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      expire_at:
        type: integer
      id:
        type: integer
      launched_at:
        type: integer
      name:
        type: string
      output:
        type: string
      params:
        items:
          type: integer
        type: array
      scheduled_started_at:
        type: integer
      updatedAt:
        type: string
    type: object
  go_shurtiner_internal_app_model.Link:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      shortened:
        type: string
      source:
        type: string
      updatedAt:
        type: string
      user:
        items:
          $ref: '#/definitions/go_shurtiner_internal_app_model.User'
        type: array
    type: object
  go_shurtiner_internal_app_model.Params:
    properties:
      name:
        type: string
      params:
        $ref: '#/definitions/json.JSON'
      scheduledStarted_at:
        type: integer
    type: object
  go_shurtiner_internal_app_model.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      lastname:
        type: string
      links:
        items:
          $ref: '#/definitions/go_shurtiner_internal_app_model.Link'
        type: array
      middlename:
        type: string
      name:
        type: string
      password:
        type: string
      updatedAt:
        type: string
    type: object
  go_shurtiner_internal_app_model.UserApi:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      lastname:
        type: string
      middlename:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  go_shurtiner_internal_http_handler.ErrorResponse:
    properties:
      details: {}
      errors:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  json.JSON:
    type: object
host: localhost:8080
info:
  contact: {}
  description: Go Shurtiner API Swagger
  title: Go Shurtiner API
  version: "1.0"
paths:
  /report:
    post:
      consumes:
      - application/json
      description: Постановка задачи генерации отчета.
      parameters:
      - description: Тело запроса
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go_shurtiner_internal_app_model.Params'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ с полной информацией о добавленных ссылках
          schema:
            items:
              $ref: '#/definitions/go_shurtiner_internal_app_model.JobQueue'
            type: array
        "400":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/go_shurtiner_internal_http_handler.ErrorResponse'
      summary: Создать отложенную задачу генерации отчета.
      tags:
      - link
  /short:
    post:
      consumes:
      - application/json
      description: Сокращение ссылок из массива параметров.
      parameters:
      - description: Тело запроса
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/go_shurtiner_internal_app_model.CreateLink'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ с полной информацией о добавленных ссылках
          schema:
            items:
              $ref: '#/definitions/go_shurtiner_internal_app_model.Link'
            type: array
        "400":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/go_shurtiner_internal_http_handler.ErrorResponse'
      summary: Сократить ссылки от имени зарегистрированного пользователя
      tags:
      - link
  /short/{link}:
    get:
      consumes:
      - application/json
      description: Поиск оригинальной ссылки по короткому идентификатору.
      parameters:
      - description: Идентификатор сокращённой ссылки
        in: path
        name: link
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ с полной информацией о ссылке
          schema:
            $ref: '#/definitions/go_shurtiner_internal_app_model.Link'
        "400":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/go_shurtiner_internal_http_handler.ErrorResponse'
      summary: Получить информацию по сокращённой ссылке
      tags:
      - link
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Информация по зарегистрированному пользователю.
      parameters:
      - description: Смещение пагинации
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ с полной информацией о пользователях
          schema:
            $ref: '#/definitions/go_shurtiner_internal_app_model.UserApi'
        "400":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/go_shurtiner_internal_http_handler.ErrorResponse'
      summary: Получить информацию по зарегистрированному пользователю.
      tags:
      - user
  /users:
    get:
      consumes:
      - application/json
      description: Информация по всем зарегистрированным пользователям.
      parameters:
      - description: Смещение пагинации
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ с полной информацией о пользователях
          schema:
            items:
              $ref: '#/definitions/go_shurtiner_internal_app_model.UserApi'
            type: array
        "400":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/go_shurtiner_internal_http_handler.ErrorResponse'
      summary: Получить информацию по зарегистрированным пользователям.
      tags:
      - user
schemes:
- http
swagger: "2.0"